Moved tests to __tests__
