+++ package.json
@@ -144,6 +144,7 @@
     }
   },
   "jest": {
+    "verbose": true,
     "testEnvironment": "node",
     "timers": "fake",
     "clearMocks": true,
@@ -166,7 +167,8 @@
     },
     "testMatch": [
       "<rootDir>/**/*.spec.ts",
-      "<rootDir>/**/*.test.ts"
+      "<rootDir>/**/*.test.ts",
+      "**/__tests__/**/*.(test|spec).ts"
     ],
     "setupFiles": [
       "./server/setupJest.ts"
